Move pod topology spread predicate logic to its filter plugin
